% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise.R
\name{add_noise}
\alias{add_noise}
\title{Add Noise to the Template}
\usage{
add_noise(
  qn,
  template_q,
  time = NULL,
  cv = 0,
  is_constrained = NULL,
  use_verbose = FALSE
)
}
\arguments{
\item{qn}{array of dimension \eqn{L \times M \times N} : contains the set of \eqn{N}
functions for which \code{template_q} is the template. Functions are of the same
dimension \eqn{L} and evaluated at the same points \eqn{M} as \code{template_q}}

\item{template_q}{either array of dimension \eqn{L \times M \times 1} or matrix
of dimension \eqn{L \times M} : contains the template function to be perrurbed
by adding noise. The function is of dimension \eqn{L} and is evaluated in \eqn{M} points.}

\item{time}{vector of dimension \eqn{M} : specifies the grid on which the functions
are evaluated. If \code{NULL}, it is set as an uniform grid in \eqn{[0,1]}.}

\item{cv}{scalar, only non-negative values are accepted : coefficient of variation (see Details).}

\item{is_constrained}{vector of size \eqn{L} : each element indicates
whether the dimension \eqn{l = 1,\dots,L} is constrained and how (see Details)}

\item{use_verbose}{boolean : specifying whether to display information about
the calculations in the console. Defaults to \code{FALSE}.}
}
\value{
\itemize{
\item \code{template_q_new} matrix of dimensions \eqn{L \times M} : original function
\code{template_q} perturbed by noise.
}
}
\description{
Generate noise with zero mean and covariance function, and use it to perturb
the given template function
}
\details{
This function adds noise to the given template function \code{template_q}.

The noise function is sampled from a gaussian distribution with null mean function
and exponential covariance function. The hyperparameter of the covariance function
are estimated by minimizing the squared Frobenius distance to the sample covariance
function of the data in \code{qn}.

By setting the coefficient of variation \code{cv}, users can fix the amount of noise to be applied.
Specifically, the amplitude of the exponential covariance function would be
proportional to the squared norm of the \code{template_q} with proportionality constant
\code{cv}. If \code{cv} is not set, the optimal amplitude is used.

\code{add_noise()} handles the case in which dimensions of the function are constrained
to be non-negative (or positive) by transforming them though a \code{sqrt} transformation
(or \code{log} transformation). Such dimensions must be specified in the parameter
\code{is_positive}

\code{is_constrained} is used to set constraints for the values of the synthetic
functions. For each dimension of the function, it states whether values are
(strictly) positive. Accepted values are: \code{pos} (positive functions),
\code{strict-pos} (non-negative functions), and \code{real} (non-constrained functions).
Constraied dimensions are transformed through sqrt-transformation (if positive
functions) or log-transformation (strict-positive functions)
}
