% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function__compute_elastic_distance.R
\name{compute_elastic_distance_N}
\alias{compute_elastic_distance_N}
\title{Elastic distances computation between a set of functions}
\usage{
compute_elastic_distance_N(f_array, time)
}
\arguments{
\item{f_array}{an array of dimensions \eqn{L \times M \times N} : contains the
values of \eqn{N} functions of dimension \eqn{L} observed over \eqn{M} points
in the domain}

\item{time}{\eqn{M}-dimensional vector, sample points of functions}
}
\value{
list of two elements:
\itemize{
\item \code{Dx_tot} matrix \eqn{N \times N} : phase distance
\item \code{Dy_tot} matrix \eqn{N \times N} : amplitude distance
}
}
\description{
Calculates pairwise the amplitude distance and the phase distance of a set of
functions. Iteratively calls the function elastic.distance.ari.
}
