% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elastic_distance.R
\name{compute_elastic_distance_one_set}
\alias{compute_elastic_distance_one_set}
\title{Elastic distances computation between a set of functions}
\usage{
compute_elastic_distance_one_set(f_array, time)
}
\arguments{
\item{f_array}{An array of dimensions \eqn{L \times M \times N} : contains
the values of \eqn{N} functions of dimension \eqn{L} observed over \eqn{M}
points in the domain.}

\item{time}{An \eqn{M}-dimensional vector, sample points of functions.}
}
\value{
A list of two elements:
\itemize{
\item \code{Dx_tot} matrix \eqn{N \times N} : phase distance
\item \code{Dy_tot} matrix \eqn{N \times N} : amplitude distance
For each matrix, the element in position \verb{[i,j]} represents the distance
between the \code{i}-th and the \code{j}-th function in \code{f_array}.
}
}
\description{
Calculates pairwise the amplitude distance and the phase distance of a set of
functions. Iteratively calls the function \code{compute_elastic_distance()}.
}
