% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elastic_distance.R
\name{compute_elastic_distance_two_sets}
\alias{compute_elastic_distance_two_sets}
\title{Elastic distances computation between two set of functions}
\usage{
compute_elastic_distance_two_sets(f_array1, f_array2, time)
}
\arguments{
\item{f_array1}{An array of dimensions \eqn{L \times M \times N1} : contains
the values of \eqn{N1} functions of dimension \eqn{L} observed over \eqn{M}
points.}

\item{f_array2}{An array of dimensions \eqn{L \times M \times N2} : contains
the values of \eqn{N2} functions of dimension \eqn{L} observed over \eqn{M}
points in the domain.}

\item{time}{An \eqn{M}-dimensional vector, sample points of functions.}
}
\value{
A list of two elements:
\itemize{
\item \code{Dx_tot} matrix \eqn{N1 \times N2} : phase distance
\item \code{Dy_tot} matrix \eqn{N1 \times N2} : amplitude distance
For each matrix, the element in position \verb{[i,j]} represents the distance
between the \code{i}-th function in \code{f_array1} and the \code{j}-th function in
\code{f_array_2}.
}
}
\description{
Calculates pairwise the amplitude distance and the phase distance of the
items in a set of functions with the items in another set of functions.
Iteratively calls the function \code{compute_elastic_distance()}.
}
