% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_optimal_w_a}
\alias{estimate_optimal_w_a}
\title{Estimate w_a}
\usage{
estimate_optimal_w_a(D1, D2, n_iterations = 5)
}
\arguments{
\item{D1}{matrix of dimension \eqn{N \times N} : first distance matrix}

\item{D2}{matrix of dimension \eqn{N \times N} : second distance matrix}

\item{n_iterations}{integer in \eqn{\{1,\dots,N-1\}} : number of iteration for
the choice of alpha. Default is set to 5.}
}
\value{
\code{w_a} : scalar, with \eqn{w_a \in [0,1]} : estimated optimal parameter.
}
\description{
Estimate the optimal value of \code{w_a} for convex combination of two distances.
}
\details{
This functions finds the value of \code{w_a} to combine two distance matrices
\code{D1} and \code{D2}. The optimal value is set as the one maximizing the
compromise between hierarchical clustering output with \code{D1} and with \code{D2}.
It follows from the algorithm \code{hclustcompro} in the library \code{SPARTAAS}.
}
