% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_weighted_karcher_mean.R
\name{multivariate_weighted_karcher_mean}
\alias{multivariate_weighted_karcher_mean}
\title{Weighted Karcher Mean of Multivariate Functional Data}
\usage{
multivariate_weighted_karcher_mean(
  beta,
  wts = NULL,
  lambda = 0,
  maxit = 20,
  ncores = 1L
)
}
\arguments{
\item{beta}{An array of sizes \eqn{L \times M \times N} and it is
interpreted as a sample of \eqn{N} \eqn{L}-dimensional curves observed on a
grid of size \eqn{M}.}

\item{wts}{Vector of length \eqn{N} with weights to be assigned to each
curve. Use \code{NULL} for unweighted mean. Defaults to \code{NULL}.}

\item{lambda}{A numeric value specifying the elasticity. Defaults to \code{0.0}.}

\item{maxit}{Maximum number of iterations.}

\item{ncores}{XXX}
}
\value{
Returns a list containing \item{qmu}{mean srvf}
\item{betamean}{(Weighted) mean curve}
\item{qmu}{(Weighted) mean srvf}
\item{beta}{Array of original curves}
\item{q}{Array of original srvfs}
\item{betan}{Array of aligned curves}
\item{qn}{Array of aligned srvfs}
\item{wts}{Weights used in the computation}
\item{E}{Energy, as the \eqn{L^2}-norm of the mean at each iteration}
\item{qun}{Cost function, as the sum of distances to the mean at each iteration}
\item{type}{String indicating whether mean or weighted mean is returned}
}
\description{
Calculates (weighted) Karcher mean of a collection of multivariate functional
data using the elastic square-root velocity (srvf) framework.
}
\references{
Srivastava, A., Klassen, E., Joshi, S., Jermyn, I., (2011). Shape
analysis of elastic curves in euclidean spaces. Pattern Analysis and Machine
Intelligence, IEEE Transactions on 33 (7), 1415-1428.
}
\keyword{alignment}
\keyword{srvf}
