% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random sample from a Dirichlet distribution}
\usage{
rdirichlet(n, alpha, method = c("beta", "gamma"))
}
\arguments{
\item{n}{integer : number of items to be sampled}

\item{alpha}{scalar, with \eqn{\alpha > 0} : concentration parameter of the
distribution}

\item{method}{method for sampling. Either \code{"beta"} or \code{"gamma"}.}
}
\value{
A numeric matrix of shape \eqn{n \times \mathrm{length}(\alpha)}
storing an $n$-sample drawn from a Dirichlet distribution with parameters
\eqn{\alpha}.
}
\description{
Random sample from a Dirichlet distribution
}
