% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvf_align.R
\name{srvf_align}
\alias{srvf_align}
\title{Alignment}
\usage{
srvf_align(
  f,
  time,
  wgt = NULL,
  seeds = NULL,
  centroid_type = c("mean", "medoid"),
  lambda = 0,
  smooth_data = FALSE,
  sparam = 25L,
  parallel = FALSE,
  alignment = TRUE,
  omethod = c("DP", "DP2", "RBFGS"),
  use_verbose = FALSE
)
}
\arguments{
\item{f}{Either a numeric matrix or a numeric 3D array specifying the
functions that need to be jointly clustered and aligned.
\itemize{
\item If a matrix, it must be of shape \eqn{M \times N}. In this case, it is
interpreted as a sample of \eqn{N} curves observed on a grid of size
\eqn{M}.
\item If a 3D array, it must be of shape \eqn{L \times M \times N} and it is
interpreted as a sample of \eqn{N} \eqn{L}-dimensional curves observed on a
grid of size \eqn{M}.
}}

\item{time}{A numeric vector of length \eqn{M} specifying the grid on which
the curves are evaluated.}

\item{wgt}{A numeric vector of length \eqn{N} specifying the weights for
each curve to be considered during the averaging process. If NULL, then
all curves are set to have the same weight = 1.}

\item{seeds}{An integer vector of length \code{K} specifying the indices of the
curves in \code{f} which will be chosen as initial centroids. Defaults to \code{NULL}
in which case such indices are randomly chosen.}

\item{centroid_type}{A string specifying the type of centroid to compute.
Choices are \code{"mean"} or \code{"medoid"}. Defaults to \code{"mean"}.}

\item{lambda}{A numeric value specifying the elasticity. Defaults to \code{0.0}.}

\item{smooth_data}{A boolean specifying whether to smooth data using a box
filter. Defaults to \code{FALSE}.}

\item{sparam}{An integer value specifying the number of box filters applied.
Defaults to \code{25L}.}

\item{parallel}{A boolean specifying whether parallel mode (using
\code{\link[foreach:foreach]{foreach::foreach()}} and the \strong{doParallel} package) should be activated.
Defaults to \code{FALSE}.}

\item{alignment}{A boolean specifying whether to perform alignment. Defaults
to \code{TRUE}.}

\item{omethod}{A string specifying which method should be used to solve the
optimization problem that provides estimated warping functions. Choices are
\code{"DP"}, \code{"DP2"} or \code{"RBFGS"}. Defaults to \code{"DP"}.}

\item{use_verbose}{A boolean specifying whether to display information about
the calculations in the console. Defaults to \code{FALSE}.}
}
\value{
An object of class \code{fdakma} which is a list containing:
\itemize{
\item \code{f0}: the original functions;
\item \code{q0}: the original SRSFs;
\item \code{fn}: the aligned functions as matrices or a 3D arrays of the same shape
than \code{f0} by clusters in a list;
\item \code{qn}: the aligned SRSFs as matrices or a 3D arrays of the same shape
than \code{f0} separated in clusters in a list;
\item \code{template}: the centroids in the original functional space;
\item \code{template.q}: the centroids in SRSF space;
\item \code{template_start} : the initial point \code{template(0)}
\item \code{gam}: the warping functions as matrices or a 3D arrays of the same shape
than \code{f0} by clusters in a list;
\item \code{qun}: cost function value.
}
}
\description{
This function aligns functions using the elastic square-root
slope function (SRSF) framework.
}
