% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function__compute_elastic_distance.R
\name{compute_elastic_distance}
\alias{compute_elastic_distance}
\title{Elastic distances computation for 2 functions}
\usage{
compute_elastic_distance(f1, f2, time, lambda = 0, pen = "roughness")
}
\arguments{
\item{f1}{\eqn{M}-dimensional vector (if unidimensional) or  \eqn{L \times M} matrix (if multidimensional)
with the \eqn{M} measurements in the \eqn{L} dimensions of the function \code{f1}}

\item{f2}{\eqn{M}-dimensional vector (if unidimensional) or a \eqn{L \times M} matrix (if multidimensional)
with the \eqn{M} measurements in the \eqn{L} dimensions of the function \code{f2}}

\item{time}{\eqn{M}-dimensional vector, sample points of functions}

\item{lambda}{numeric in \eqn{[0,1]}, controlling the amount of warping}

\item{pen}{alignment penalty, between "roughness" (second derivative), "geodesic"
(geodesic distance from id), "norm" (norm from id)}
}
\value{
list of two elements:
\itemize{
\item \code{Dx} scalar : phase distance
\item \code{Dy} scalar : amplitude distance
}
}
\description{
Calculates the amplitude distance and the phase distance of two functions
}
