% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_weighted_mean_inverse.R
\name{sqrt_weighted_mean_inverse}
\alias{sqrt_weighted_mean_inverse}
\title{SRVF transform of warping functions}
\usage{
sqrt_weighted_mean_inverse(gam, wts = NULL)
}
\arguments{
\item{gam}{matrix (\eqn{N} x \eqn{M}) of \eqn{M} warping functions with
\eqn{N} samples}

\item{wts}{vector of weights}
}
\value{
\code{gamI} inverse of (weighted) Karcher mean warping function
}
\description{
This function calculates the srvf of warping functions with corresponding
shooting vectors and finds the inverse of (weighted) mean.
}
\details{
It is based on the \code{SqrtMeanInverse} funcion of the \code{fdasrvf} package, and
enlarges it potentiality by computing the mean both non-weighted and weighted
}
\examples{
gamI <- sqrt_weighted_mean_inverse(fdasrvf::simu_warp$warping_functions)
}
\references{
Srivastava, A., Wu, W., Kurtek, S., Klassen, E., Marron, J. S.,
May 2011. Registration of functional data using fisher-rao metric,
arXiv:1103.3817v2.

Tucker, J. D., Wu, W., Srivastava, A.,
Generative Models for Function Data using Phase and Amplitude Separation,
Computational Statistics and Data Analysis (2012), 10.1016/j.csda.2012.12.001.
}
\keyword{alignment}
\keyword{srvf}
